pyproject.toml
[project]
name = "parallax"
version = "0.1.0"
description = "Embedding-based biosecurity screening for AI-redesigned proteins"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.86",
    "fair-esm>=2.0.0",
    "fastapi[standard]>=0.135.1",
    "numpy>=2.4.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "uvicorn>=0.41.0",
]
